vertex-ai-data-store-loader/requirements.txt
google-cloud-bigquery
google-cloud-storage
google-cloud-discoveryengine
python-dotenv